# Services/Server/song-content/openapi.yaml
openapi: 3.1.0
info:
  title: Mono Song Content API
  version: 1.0.0
servers:
  - url: https://ncm.zijiu522.cn
security:
  - appToken: []
paths:
...
  /api/song-content/content/{versionId}:
    get:
      summary: 获取审核上下文、字段差异、来源和生成成本
      security: [{ adminToken: [] }]
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 内容审核数据 }
  /api/song-content/content/{versionId}/publish:
    post:
      summary: 原子发布内容版本
      security: [{ adminToken: [] }]
      responses:
        '200': { description: 已发布 }
...
components:
  securitySchemes:
    appToken:
      type: apiKey
      in: query
      name: token
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
...